GP2-3227: Basic laying in the atlas investment opps listing page
This is based on the capinvest code and will be extended with extra filters as required. The current filters are ported over, including changing behaviour to checking against multiple regions, not just one.

Code has largely been copied fro capinvest code, and that code will be removed in a later commit