GP2-3221: Update International Home Page to accept new, cut-down data from API
This changeset strips out most of the old markup from the international homepage and lays in some skeleton data-only content ahead of new styling.

It is designed to consume content provided by the CMS once https://github.com/uktrade/directory-cms/pull/901 has been merged to that service.

Note that for now the API data does not reference the hero image or CSS background colour for the page, which will be hard-coded for MVP.

Many tests for the homepage content have been removed as they would otherwise be failing due to the absence of the older content

Also, note that there was a 'new_international_landing_page.html' which appears NOT to be the one being used, so I've added a comment at the top of it to make this clearer for the future. It can likely be deleted.