Use squoosh for image optimisation instead of imagemin.
The "imagemin" package and its dependency the "download" package are both
unmaintained packages. The "download" package has a vulnerable
dependency on the "got" package, which we can't reasonably expect
to be patched due to the package being unmaintained.

This change updates the webpack config to use the squoosh minimiser
instead of the imagemin minimiser to avoid this vulnerable dependency.